requirements.txt
azure-eventhub
streamlit
python-dotenv